Add debug support back to rephract
	-Drephract.debug=NOTNULL
	-Drephract.debug.threshold={INTEGER|0}

If rephract.debug is set to non-null, then use the threshold (defaults to 0)
to determine if information regarding link guard-testing is printed.

Move the link-invocation logic from RephractLinker to the LinkPlan.

Synchronize it (not as bad as it sounds).

Avoid duplicate race-condition linking that allows for deep/duplicitous links
in a single call-site due to multiple threads linking at the same place and same
time.